Move net settlement group to market metering point
